Fixed tests

+++ database/migrations/2020_08_26_110918_create_tasks_table.php
@@ -17,15 +17,15 @@ class CreateTasksTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('description');
-            $table->unsignedInteger('status_id');
+            $table->bigInteger('status_id');
             $table->foreign('status_id')
                 ->references('id')
                 ->on('statuses');
-            $table->unsignedInteger('created_by_id');
+            $table->bigInteger('created_by_id');
             $table->foreign('created_by_id')
                 ->references('id')
                 ->on('users');
-            $table->unsignedInteger('assigned_to_id');
+            $table->bigInteger('assigned_to_id');
             $table->foreign('assigned_to_id')
                 ->references('id')
                 ->on('users');
